Remove some NOEXCEPT from pack interface
MSVC2013 is flaky when forceinline/noexcept are
combined when the noexcept function is actually
not noexcept because internal calls are not.

Using a expression based noecept works on MSVC2015
but not on MSVC2013. This patch remove the offenders
as currently performance of force inlined function
are more important than actual noexcept there.

This has to be amended later.