Make storage_of aware of non-supported types
Two cases were invalid in the case where hardware
has incompete type support:

On some extension, even if the expected cardinal we
compute may make sense, the extension traits may
tell us a given type is not supported. In this case,
we need to find one elsewhere.

On some extension, no register types exists for
storing N elements of type T. In this case, we
trigger the search or the best emulation possible.