Fix evaluation order in load
The "use ... to do our bidding" was partly faulty as
some ... constructs does not guarantee sub function calls
orders. The old code was working as expected on gcc/clang
but MSVC and mostly ICC was playing games and evaluated
the underlying *p++ in reverse order, leading to failures.

The code now use a local initializer_list and make to
be sure everything is in order and in *the correct* order.