Aggregate storage is now more idiomatic
The storage of aggregation-based emulated pack was using
an arbitrary number o sub-vector. This makes a lot of
function needlessly complex to write as we can't know
for sure that the number of sub-element between two
different aggregated pack are equal.

This patch makes this emulation more in line with x86
style SIMD types by makign each aggregation to be
exactly a pair of sub-pack.