Implement slide
slide<N>(a,b) perform an element per element shift of value across multiple
packs. The offset N goes from -cardinal_of(a) to +cardinal_of(a) and represent
the amount of element to shift and the ordering of the shift with the following
rule :

  slide<-K>(a,b) == slide<cardinal_of(a) - K>(b,a)

slide is primarily used to implement aligned_load with static misalignment
and some other functions where positional shift are required.

                      ****BREAKING CHANGE******
Support for offset greater than the pack cardinal has been removed as it
proved complicated the implementation for no actual benefit for users.