Introduce simd::segmented_output_range
Takes a Range (as pair of iterators or actual range) and
turn it into a triplet of range describing the potential
prologue, main SIMD segment and epilogue of the original
Range. This "trichotomoy" is then usable to apply operations
on the different segment of the initial range - the first
non-SIMD segment, the SIMD segment and the final non-SIMD
segment- without to use complicated computations of boundaries.