Implement VMX shuffle
Like for SSSE3 (or the other way around if you go
by history of hardware), we use a index-to-byte
conversion to build the permutation control vector.

Zeroing is handled as a splat+bitwise_and in the
binary case but optimized as a binary shuffle with
zero in the unary case to save a very costly splat
of arbitrary values.