Remove warning in pack-to-storage_type conversion
Returning the underlying storage of a pack is done
without any consideration of previous intialization
to limit unexpected code generation. G++ was wary
of such a piece of code and warned about using said
storage in an unitiliazed state.

This patch use GCC diagnostic pragmas to locally
silence this warning as the code is purposedly
written with this behaviour in mind.