Provide a more sensible alias over aligned_alloc
Instead of just forwarding align::aligned_alloc, we
provide a typed boost::simd::allocate that use the
type and architectural informations to select the
correct alignment.

Size of the allocation is also simpler as it is passed
in number of elements instead o number of byte.

Lastly, documentation has been amended for those new
functions.