[ABI Change] frexp extensions (#339)
frexp behavior is changed to return a pair of flaoting point value.
ifrexp is added to access the old frexp behavior of returning a pair of real/integer for the mantissa/exponent.

Both function has been given a full raw/pedantic support to make function using them easier to optimize. Now frepx/ifrexp behavior is consistent with the old "fast" version. pedantic decorator is to be used to get the perfectly standard behavior