MSVC: BOOST_FORCEINLINE -> BOOST_MAYBEINLINE
This fixes the crazy compilation/linking time for some tests with MSVC. The forceinline has been
replaced by a simple inline. This should not really impact performances because most of those
functions have a big computation time. Also, this is not possible to live with that as it takes
more than 20min to compile/link some tests... You can still force MAYBEINLINE to be used as
FORCEINLINE with BOOST_SIMD_NO_MAYBEINLINE but you will get back to those crazy times...

Close #479